--- ToolGood.ReadyGo3/DataCentxt/Enums/ColumnType.cs
@@ -1,29 +1,0 @@
-﻿namespace ToolGood.ReadyGo3.DataCentxt.Enums
-{
-    /// <summary>
-    /// 列类型 
-    /// </summary>
-    public enum ColumnType
-    {
-        /// <summary>
-        /// 未设置
-        /// </summary>
-        None,
-        /// <summary>
-        /// 列
-        /// </summary>
-        Column = 1,
-        /// <summary>
-        /// SQL代码
-        /// </summary>
-        Code = 2,
-        /// <summary>
-        /// 函数
-        /// </summary>
-        Function = 4,
-        /// <summary>
-        /// 值
-        /// </summary>
-        Value = 8
-    }
-}
