--- ToolGood.ReadyGo3/DataCentxt/Enums/WhereType.cs
@@ -1,25 +1,0 @@
-﻿namespace ToolGood.ReadyGo3.DataCentxt.Enums
-{
-    /// <summary>
-    /// Where类型
-    /// </summary>
-    public enum WhereType
-    {
-        /// <summary>
-        /// 空
-        /// </summary>
-        None,
-        /// <summary>
-        /// 单个
-        /// </summary>
-        Single,
-        /// <summary>
-        /// 并且
-        /// </summary>
-        And,
-        /// <summary>
-        /// 或
-        /// </summary>
-        Or
-    }
-}
