--- ToolGood.ReadyGo3/DataCentxt/Exceptions/NoJoinOnException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace ToolGood.ReadyGo3.DataCentxt.Exceptions
-{
-    /// <summary>
-    /// 没有Join On sql代码异常
-    /// </summary>
-    public class NoJoinOnException : Exception
-    {
-
-
-
-
-    }
-}
