--- ToolGood.ReadyGo3/DataCentxt/Exceptions/NoPrimaryKeyException.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace ToolGood.ReadyGo3.DataCentxt.Exceptions
-{
-    /// <summary>
-    /// 没有主键异常
-    /// </summary>
-    public class NoPrimaryKeyException : Exception
-    {
-
-
-    }
-}
