--- ToolGood.ReadyGo3/DataCentxt/Exceptions/NoWhereException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace ToolGood.ReadyGo3.DataCentxt.Exceptions
-{
-    /// <summary>
-    /// 没有WHERE异常
-    /// </summary>
-    public class NoWhereException : Exception
-    {
-    }
-}
