--- ToolGood.ReadyGo3/DataCentxt/QTable.cs
@@ -1,232 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-using ToolGood.ReadyGo3.DataCentxt.Internals;
-using ToolGood.ReadyGo3.Gadget.Internals;
-
-namespace ToolGood.ReadyGo3.DataCentxt
-{
-    public abstract partial class QTable : IDisposable
-    {
- 
-        /// <summary>
-        /// SQL记录
-        /// </summary>
-        public SqlRecord __SQL__ { get { return GetSqlHelper()._Sql; } }
-        private readonly SqlHelper _sqlHelper;
-
-        internal SqlBuilder _sqlBuilder;
-        internal string _asName;
-        internal JoinType _joinType;
-        internal QJoinCondition _joinCondition;
-        internal Dictionary<string, QTableColumn> _columns = new Dictionary<string, QTableColumn>();
-        internal QTableColumn _primaryKey;
-        /// <summary>
-        /// 
-        /// </summary>
-        protected QTable() {                    }
-
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="sqlHelper"></param>
-        protected QTable(SqlHelper sqlHelper)
-        {
-            _sqlHelper = sqlHelper;
-        }
-        /// <summary>
-        /// 
-        /// </summary>
-        public void Dispose()
-        {
-         
-        }
-
-        #region AddColumn
-        /// <summary>
-        /// 添加列
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="columnName"></param>
-        /// <param name="fieldName"></param>
-        /// <param name="isPk"></param>
-        /// <param name="isAutoIncrement"></param>
-        /// <param name="resultColumn"></param>
-        /// <param name="resultSql"></param>
-        /// <returns></returns>
-        protected QTableColumn<T> AddColumn<T>(string columnName, string fieldName, bool isPk, bool isAutoIncrement, bool resultColumn, string resultSql)
-        {
-            QTableColumn<T> column = new QTableColumn<T> {
-                _columnType = Enums.ColumnType.Column,
-                _columnName = columnName,
-                _isPrimaryKey = isPk,
-                _isResultColumn = resultColumn,
-                _resultSql = resultSql,
-                _table = this,
-                _isAutoIncrement = isAutoIncrement
-            };
-            _columns.Add(fieldName.ToLower(), column);
-            if (isPk) {
-                _primaryKey = column;
-            }
-            return column;
-        }
-        /// <summary>
-        /// 添加列
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="columnName"></param>
-        /// <param name="fieldName"></param>
-        /// <param name="isPk"></param>
-        /// <param name="isAutoIncrement"></param>
-        /// <returns></returns>
-        protected QTableColumn<T> AddColumn<T>(string columnName, string fieldName, bool isPk, bool? isAutoIncrement = null)
-        {
-            if (isPk) {
-                if (isAutoIncrement == null) {
-                    var type = typeof(T);
-                    if (type == typeof(Int16) || type == typeof(int) || type == typeof(Int64) ||
-                        type == typeof(UInt16) || type == typeof(UInt32) || type == typeof(UInt64)) {
-                        return AddColumn<T>(columnName, fieldName, isPk, true, false, null);
-                    }
-                    return AddColumn<T>(columnName, fieldName, isPk, false, false, null);
-                }
-                return AddColumn<T>(columnName, fieldName, isPk, isAutoIncrement.Value, false, null);
-            }
-            return AddColumn<T>(columnName, fieldName, isPk, false, false, null);
-        }
-        /// <summary>
-        /// 添加列
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="columnName"></param>
-        /// <param name="fieldName"></param>
-        /// <param name="resultSql"></param>
-        /// <returns></returns>
-        protected QTableColumn<T> AddColumn<T>(string columnName, string fieldName, string resultSql)
-        {
-            if (string.IsNullOrEmpty(resultSql) == false) {
-                resultSql = resultSql.Replace("{0}.", "{0}");
-                if (resultSql[0] != '(') {
-                    resultSql = "(" + resultSql + ")";
-                }
-            }
-            return AddColumn<T>(columnName, fieldName, false, false, true, resultSql);
-        }
-
-
-        #endregion
-
-        /// <summary>
-        /// 清空，进行下一次
-        /// </summary>
-        public void Clear()
-        {
-            if (_sqlBuilder != null) {
-                _sqlBuilder.Dispose();
-                _sqlBuilder = null;
-            }
-        }
-
-
-        /// <summary>
-        /// 
-        /// </summary>
-        protected abstract void Init();
-
-        internal SqlHelper GetSqlHelper()
-        {
-            //            if (_sqlHelper == null || _sqlHelper._isDisposable) {
-            //#if !NETSTANDARD2_0
-
-            //                if (string.IsNullOrEmpty(_connStringName) == false) {
-            //                    _singleSqlHelper = true;
-            //                    _sqlHelper = new SqlHelper(_connStringName);
-            //                } else if (SqlHelper._lastSqlHelper != null && SqlHelper._lastSqlHelper._isDisposable == false) {
-            //                    _singleSqlHelper = false;
-            //                    _sqlHelper = SqlHelper._lastSqlHelper;
-            //                } else {
-            //                    _singleSqlHelper = true;
-            //                    _sqlHelper = new SqlHelper();
-            //                }
-            //#else
-            //                            if (SqlHelper._lastSqlHelper != null && SqlHelper._lastSqlHelper._isDisposable == false) {
-            //                                _singleSqlHelper = false;
-            //                                _sqlHelper = SqlHelper._lastSqlHelper;
-            //                            }
-            //                            throw new InvalidProgramException();
-            //#endif
-            //            }
-            return _sqlHelper;
-        }
-
-        internal SqlType GetSqlType()
-        {
-            return GetSqlHelper()._sqlType;
-        }
-        internal SqlBuilder GetSqlBuilder()
-        {
-            if (_sqlBuilder == null) {
-                _sqlBuilder = new SqlBuilder(this);
-            }
-            return _sqlBuilder;
-        }
-        internal virtual Type GetTableType()
-        {
-            return null;
-        }
-
-    }
-
-    public abstract partial class QTable<T> : QTable
-        where T : class
-    {
-        /// <summary>
-        /// 
-        /// </summary>
-        protected QTable() : base()
-        {
-            Init();
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="sqlHelper"></param>
-        protected QTable(SqlHelper sqlHelper) : base(sqlHelper)
-        {
-            Init();
-        }
-
-
-        /// <summary>
-        /// 添加列
-        /// </summary>
-        /// <typeparam name="T1"></typeparam>
-        /// <param name="fieldName"></param>
-        /// <returns></returns>
-        protected QTableColumn<T1> AddColumn<T1>(string fieldName)
-        {
-            var pd = PetaPoco.Core.PocoData.ForType(typeof(T));
-            var column = pd.Columns.Where(q => q.Value.PropertyName == fieldName).Select(q => q.Value).FirstOrDefault();
-            if (column != null) {
-                var isPk = pd.TableInfo.PrimaryKey == fieldName;
-                return AddColumn<T1>(column.ColumnName, fieldName, isPk, isPk ? pd.TableInfo.AutoIncrement : false, column.ResultColumn, column.ResultSql);
-            }
-            return new QTableColumn<T1>();
-        }
-
-        private Type _type;
-        internal override Type GetTableType()
-        {
-            if (_type==null) {
-                _type= typeof(T);
-            }
-            return _type;
-        }
-    }
-
-
-}
