--- ToolGood.ReadyGo3/Enums/JoinType.cs
@@ -1,26 +1,0 @@
-﻿namespace ToolGood.ReadyGo3
-{
-    /// <summary>
-    /// SQL JOIN类型
-    /// </summary>
-    public enum JoinType
-    {
-        /// <summary>
-        /// Inner Join
-        /// </summary>
-        Inner,
-        /// <summary>
-        /// Left Join
-        /// </summary>
-        Left,
-        /// <summary>
-        /// Right Join
-        /// </summary>
-        Right,
-        /// <summary>
-        /// Full Join
-        /// </summary>
-        Full
-    }
-    
-}
