+++ ToolGood.ReadyGo3/Exceptions/SqlExecuteException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ToolGood.ReadyGo3.DataCentxt.Exceptions
+namespace ToolGood.ReadyGo3.Exceptions
 {
     /// <summary>
     /// SQL执行异常
