+++ ToolGood.ReadyGo3/PetaPoco/Providers/FirebirdDbDatabaseProvider.cs
@@ -2,6 +2,7 @@
 using System.Data.Common;
 using System.Linq;
 using System.Text;
+using ToolGood.ReadyGo3.Enums;
 using ToolGood.ReadyGo3.PetaPoco.Core;
 using ToolGood.ReadyGo3.PetaPoco.Utilities;
 
@@ -9,6 +10,14 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
 {
     public class FirebirdDbDatabaseProvider : DatabaseProvider
     {
+        /// <summary>
+        /// Firebird
+        /// </summary>
+        public FirebirdDbDatabaseProvider()
+        {
+            usedEscapeSql = true;
+            escapeSql = '"';
+        }
         public override DbProviderFactory GetFactory()
         {
             return GetFactory(
@@ -80,6 +89,47 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
             }
             sb.AppendFormat($" ROWS {offset + 1} TO {offset + limit}");
             return sb.ToString();
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="function"></param>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public override string CreateFunction(SqlFunction function, params object[] args)
+        {
+            //  http://www.firebirdsql.org/refdocs/langrefupd21.html
+            switch (function) {
+                case SqlFunction.Fuction: break;
+                case SqlFunction.Len: break;
+                case SqlFunction.Max: break;
+                case SqlFunction.Min: break;
+                case SqlFunction.Avg: break;
+                case SqlFunction.Sum: break;
+                case SqlFunction.Count: break;
+                case SqlFunction.CountDistinct: break;
+                case SqlFunction.DatePart: break;
+                case SqlFunction.DateDiff: return CreateFunction("DATEDIFF(day,{0},{1})", args);
+                case SqlFunction.Year: return CreateFunction("EXTRACT(YEAR FROM {0})", args);
+                case SqlFunction.Month: return CreateFunction("EXTRACT(MONTH FROM {0})", args);
+                case SqlFunction.Day: return CreateFunction("EXTRACT(DAY FROM {0})", args);
+                case SqlFunction.Hour: return CreateFunction("EXTRACT(HOUR FROM {0})", args);
+                case SqlFunction.Minute: return CreateFunction("EXTRACT(MINUTE FROM {0})", args);
+                case SqlFunction.Second: return CreateFunction("EXTRACT(SECOND FROM {0})", args);
+                case SqlFunction.DayOfYear: return CreateFunction("EXTRACT(DAYOFYEAR FROM {0})", args);
+                case SqlFunction.Week: return CreateFunction("EXTRACT(WEEK FROM {0})", args);
+                case SqlFunction.WeekDay: return CreateFunction("EXTRACT(WEEKDAY FROM {0})", args);
+                case SqlFunction.SubString3: return CreateFunction("SUBSTRING({0} FROM {1} FOR {2})", args);
+                case SqlFunction.SubString2: return CreateFunction("SUBSTRING({0} FROM {1})", args);
+                case SqlFunction.Left: return CreateFunction("SLEFT({0},{1})", args);
+                case SqlFunction.Right: return CreateFunction("SRIGHT({0},{1})", args);
+                case SqlFunction.Lower: break;
+                case SqlFunction.Upper: break;
+                case SqlFunction.Ascii: break;
+                case SqlFunction.Concat: break;
+                default: break;
+            }
+            return base.CreateFunction(function, args);
         }
 
     }
