+++ ToolGood.ReadyGo3/PetaPoco/Providers/MsAccessDbDatabaseProvider.cs
@@ -2,7 +2,8 @@
 using System.Data;
 using System.Data.Common;
 using System.Text;
-using ToolGood.ReadyGo3.DataCentxt.Exceptions;
+using ToolGood.ReadyGo3.Enums;
+using ToolGood.ReadyGo3.Exceptions;
 using ToolGood.ReadyGo3.PetaPoco.Core;
 using ToolGood.ReadyGo3.PetaPoco.Utilities;
 
@@ -53,6 +54,48 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
                 return sb.ToString();
             }
             throw new DatabaseUnsupportException();
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="function"></param>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public override string CreateFunction(SqlFunction function, params object[] args)
+        {
+            switch (function) {
+                case SqlFunction.Fuction: break;
+                case SqlFunction.Len: break;
+                case SqlFunction.Max: break;
+                case SqlFunction.Min: break;
+                case SqlFunction.Avg: break;
+                case SqlFunction.Sum: break;
+                case SqlFunction.Count: break;
+                case SqlFunction.CountDistinct: break;
+                case SqlFunction.DatePart: break;
+                case SqlFunction.DateDiff: break;
+                case SqlFunction.Year: break;
+                case SqlFunction.Month: break;
+                case SqlFunction.Day: break;
+                case SqlFunction.Hour: break;
+                case SqlFunction.Minute: break;
+                case SqlFunction.Second: break;
+                case SqlFunction.DayOfYear: break;
+                case SqlFunction.Week: break;
+                case SqlFunction.WeekDay: break;
+                case SqlFunction.SubString3: return CreateFunction("MID({0}, {1}, {2})", args);
+                case SqlFunction.SubString2: return CreateFunction("MID({0}, {1})", args);
+                case SqlFunction.Left: break;
+                case SqlFunction.Right: break;
+                case SqlFunction.Lower: return CreateFunction("LCASE({0})", args);
+                case SqlFunction.Upper: return CreateFunction("UCASE({0})", args);
+                case SqlFunction.Ascii: return CreateFunction("ASC({0})", args);
+                case SqlFunction.Concat: break;
+                default: break;
+            }
+
+            return base.CreateFunction(function, args);
         }
     }
 }
