+++ ToolGood.ReadyGo3/PetaPoco/Providers/PostgreSQLDatabaseProvider.cs
@@ -1,11 +1,21 @@
 ﻿using System.Data.Common;
 using System.Text;
+using ToolGood.ReadyGo3.Enums;
 using ToolGood.ReadyGo3.PetaPoco.Core;
 
 namespace ToolGood.ReadyGo3.PetaPoco.Providers
 {
     public class PostgreSQLDatabaseProvider : DatabaseProvider
     {
+        /// <summary>
+        /// 
+        /// </summary>
+        public PostgreSQLDatabaseProvider()
+        {
+            usedEscapeSql = true;
+            escapeSql = '"';
+        }
+
         public override bool HasNativeGuidSupport
         {
             get { return true; }
@@ -88,6 +98,51 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
                 sb.Append(limit);
             }
             return sb.ToString();
+        }
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="function"></param>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public override string CreateFunction(SqlFunction function, params object[] args)
+        {
+            //  https://www.postgresql.org/docs/9.1/static/functions-math.html
+            //  https://www.postgresql.org/docs/9.1/static/functions-string.html
+            //  https://www.postgresql.org/docs/9.1/static/functions-datetime.html
+            switch (function) {
+                case SqlFunction.Fuction: break;
+                case SqlFunction.Len: return CreateFunction("LENGTH({0})", args);
+                case SqlFunction.Max: break;
+                case SqlFunction.Min: break;
+                case SqlFunction.Avg: break;
+                case SqlFunction.Sum: break;
+                case SqlFunction.Count: break;
+                case SqlFunction.CountDistinct: break;
+                case SqlFunction.DatePart: break;
+                case SqlFunction.DateDiff: return CreateFunction("AGE(TIMESTAMP {0}, TIMESTAMP {1})", args);
+                case SqlFunction.Year: return CreateFunction("EXTRACT(YEAR FROM TIMESTAMP {0})", args);
+                case SqlFunction.Month: return CreateFunction("EXTRACT(MONTH FROM TIMESTAMP {0})", args);
+                case SqlFunction.Day: return CreateFunction("EXTRACT(DAY FROM TIMESTAMP {0})", args);
+                case SqlFunction.Hour: return CreateFunction("EXTRACT(HOUR FROM TIMESTAMP {0})", args);
+                case SqlFunction.Minute: return CreateFunction("EXTRACT(MINUTE FROM TIMESTAMP {0})", args);
+                case SqlFunction.Second: return CreateFunction("EXTRACT(SECOND FROM TIMESTAMP {0})", args);
+                case SqlFunction.DayOfYear: return CreateFunction("EXTRACT(DAYOFYEAR FROM TIMESTAMP {0})", args);
+                case SqlFunction.Week: return CreateFunction("EXTRACT(WEEK FROM TIMESTAMP {0})", args);
+                case SqlFunction.WeekDay: return CreateFunction("EXTRACT(WEEKDAY FROM TIMESTAMP {0})", args);
+                case SqlFunction.SubString3: return CreateFunction("SUBSTR({0}, {1}, {2})", args);
+                case SqlFunction.SubString2: return CreateFunction("SUBSTR({0}, {1})", args);
+                case SqlFunction.Left: break;
+                case SqlFunction.Right: break;
+                case SqlFunction.Lower: break;
+                case SqlFunction.Upper: break;
+                case SqlFunction.Ascii: break;
+                case SqlFunction.Concat: break;
+                default: break;
+            }
+            return base.CreateFunction(function, args);
         }
     }
 }
