+++ ToolGood.ReadyGo3/PetaPoco/Providers/SQLiteDatabaseProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System.Data.Common;
 using System.Text;
+using ToolGood.ReadyGo3.Enums;
 using ToolGood.ReadyGo3.PetaPoco.Core;
 
 namespace ToolGood.ReadyGo3.PetaPoco.Providers
@@ -68,6 +69,50 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
                 sb.Append(limit);
             }
             return sb.ToString();
+        }
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="function"></param>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public override string CreateFunction(SqlFunction function, params object[] args)
+        {
+            // http://www.sqlite.org/lang_corefunc.html
+            switch (function) {
+                case SqlFunction.Fuction: break;
+                case SqlFunction.Len: return CreateFunction("LENGTH({0})", args);
+                case SqlFunction.Max: break;
+                case SqlFunction.Min: break;
+                case SqlFunction.Avg: break;
+                case SqlFunction.Sum: break;
+                case SqlFunction.Count: break;
+                case SqlFunction.CountDistinct: break;
+                case SqlFunction.DatePart: break;
+                case SqlFunction.DateDiff: return CreateFunction("FLOOR(CAST(STRFTIME('%J',{0}) AS INT) - CAST(STRFTIME('%J',{1}) AS INT))", args);
+                case SqlFunction.Year: return CreateFunction("CAST(STRFTIME('%Y',{0}) AS INT)", args);
+                case SqlFunction.Month: return CreateFunction("CAST(STRFTIME('%m',{0}) AS INT)", args);
+                case SqlFunction.Day: return CreateFunction("CAST(STRFTIME('%d',{0}) AS INT)", args);
+                case SqlFunction.Hour: return CreateFunction("CAST(STRFTIME('%H',{0}) AS INT)", args);
+                case SqlFunction.Minute: return CreateFunction("CAST(STRFTIME('%M',{0}) AS INT)", args);
+                case SqlFunction.Second: return CreateFunction("CAST(STRFTIME('%S',{0}) AS INT)", args);
+                case SqlFunction.DayOfYear: return CreateFunction("CAST(STRFTIME('%j',{0}) AS INT)", args);
+                case SqlFunction.Week: return CreateFunction("CAST(STRFTIME('%W',{0}) AS INT)", args);
+                case SqlFunction.WeekDay: return CreateFunction("CAST(STRFTIME('%w',{0}) AS INT)", args);
+                case SqlFunction.SubString3: return CreateFunction("SUBSTR({0},{1},{2})", args);
+                case SqlFunction.SubString2: return CreateFunction("SUBSTR({0},{1})", args);
+                case SqlFunction.Left: return CreateFunction("SUBSTR({0},1,{1})", args);
+                case SqlFunction.Right: return CreateFunction("SUBSTR({0},-{1})", args);
+                case SqlFunction.Lower: break;
+                case SqlFunction.Upper: break;
+                case SqlFunction.Ascii: break;
+                case SqlFunction.Concat: break;
+                default: break;
+            }
+
+            return base.CreateFunction(function, args);
         }
     }
 }
