+++ ToolGood.ReadyGo3/PetaPoco/Utilities/AutoSelectHelper.cs
@@ -1,8 +1,6 @@
-﻿using System.Linq;
-using System;
-using System.Text;
+﻿using System.Text;
 using System.Text.RegularExpressions;
-using ToolGood.ReadyGo3.DataCentxt.Exceptions;
+using ToolGood.ReadyGo3.Exceptions;
 using ToolGood.ReadyGo3.PetaPoco.Core;
 
 namespace ToolGood.ReadyGo3.PetaPoco.Internal
